pyproject.toml
[project]
name = "ghtest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "import-ipynb>=0.2",
    "notebook>=7.4.7",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "vcrpy>=7.0.0",
]
